Move some extensions out of their package
