+++ BotCommands-core/src/test/kotlin/io/github/freya022/botcommands/test/commands/slash/SlashEventWaiter.kt
@@ -1,6 +1,6 @@
 package io.github.freya022.botcommands.test.commands.slash
 
-import dev.freya02.botcommands.jda.ktx.messages.deletes.deleteDelayed
+import dev.freya02.botcommands.jda.ktx.messages.deleteDelayed
 import dev.freya02.botcommands.jda.ktx.messages.reply_
 import io.github.freya022.botcommands.api.commands.annotations.Command
 import io.github.freya022.botcommands.api.commands.application.ApplicationCommand
