+++ BotCommands-core/src/test/kotlin/io/github/freya022/botcommands/test/commands/slash/SlashSay.kt
@@ -1,7 +1,7 @@
 package io.github.freya022.botcommands.test.commands.slash
 
 import dev.freya02.botcommands.jda.ktx.coroutines.await
-import dev.freya02.botcommands.jda.ktx.messages.deletes.deleteDelayed
+import dev.freya02.botcommands.jda.ktx.messages.deleteDelayed
 import dev.freya02.botcommands.jda.ktx.messages.reply_
 import dev.freya02.jda.emojis.unicode.UnicodeEmojis
 import io.github.freya022.botcommands.api.commands.annotations.Command
